Split price and trade historian into their own files
All going under a common history module
