+++ rotkehlchen/history/__init__.py
@@ -1,0 +1,2 @@
+from .price import PriceHistorian  # noqa: F401
+from .trades import TradesHistorian  # noqa: F401
